feat(entities): lift step/final_step to i64 (Wave 24 BIGINT migration)
- bpb_samples::Model.step: i32 -> i64 (matches new BIGINT column)
- igla_race_trials::Model.final_step: Option<i32> -> Option<i64>
- steps_done already i64, no change needed
- Updated doc-comment in bpb_samples.rs: step INT -> step BIGINT

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Refs #118

+++ src/entities/igla_race_trials.rs
@@ -14,7 +14,7 @@ pub struct Model {
     pub agent_id: Option<String>,
     pub branch: Option<String>,
     pub final_bpb: Option<f64>,
-    pub final_step: Option<i32>,
+    pub final_step: Option<i64>,
     pub bpb_latest: Option<f64>,
     pub steps_done: Option<i64>,
     pub bpb_final: Option<f64>,
